Only store past 8 turns of draw info per player, also some cleanup for future improved stacked performance

+++ defines.h
@@ -78,7 +78,7 @@ struct gamestate {
 	/* AI to invoke for each player slot */
 	uint_fast32_t (*ai[MAXPLRS])(const struct aistate* const restrict);
 	/* A bitmap of recent drawing events, 2 bits per turn */
-	uint_fast64_t draws[MAXPLRS];
+	uint_fast16_t draws[MAXPLRS];
 };
