+++ engine.c
@@ -78,7 +78,7 @@ void initGameState(struct gamestate* const restrict gs, const size_t nplayers, u
 	gs->magic = false;
 	gs->drew = false;
 	memcpy(gs->ai, ai, nplayers * sizeof(uint_fast32_t (*const)(const struct aistate* const restrict)));
-	memset(gs->draws, 0, nplayers * sizeof(uint_fast64_t));
+	memset(gs->draws, 0, nplayers * sizeof(uint_fast16_t));
 }
 
 void cleanGameState(struct gamestate* const restrict gs)
@@ -88,7 +88,7 @@ void cleanGameState(struct gamestate* const restrict gs)
 	gs->pile.top = gs->pile.c;
 	gs->deck.n = 0;
 	gs->pile.n = 0;
-	memset(gs->draws, 0, gs->nplayers * sizeof(uint_fast64_t));
+	memset(gs->draws, 0, gs->nplayers * sizeof(uint_fast16_t));
 }
 
 __attribute__((nonnull,hot)) static void removeCard(struct player* const restrict player, const card_t c)
